refactor(linker): move LinkerOutput to marked_up_text_chunk.py and make a factory method for creating MUTCs based on debug flag
